fixed #59, tests updated. Toolchain rebuild required

+++ lib/libc/rmdir.c
@@ -33,16 +33,20 @@
 
 int zrt_zcall_rmdir(const char *pathname){
     CHECK_EXIT_IF_ZRT_NOT_READY;
-
+    char* absolute_path;
+    char temp_path[PATH_MAX];
+    int ret=-1;
     LOG_SYSCALL_START("pathname=%s", pathname);
     VALIDATE_SUBSTITUTED_SYSCALL_PTR(pathname);
     errno=0;
 
     struct MountsInterface* transpar_mount = transparent_mount();
     assert(transpar_mount);
-    char temp_path[PATH_MAX];
-    char* absolute_path = zrealpath( pathname, temp_path );
-    int ret = transpar_mount->rmdir( absolute_path );
+
+    if ( (absolute_path = realpath(pathname, temp_path)) != NULL ){
+	ret = transpar_mount->rmdir( absolute_path );
+    }
+
     LOG_SHORT_SYSCALL_FINISH(ret, "pathname=%s", pathname);
     return ret;
 }
